Make scorecard generic
