+++ packages/primary/src/js/main.ts
@@ -6,6 +6,7 @@ import {
   PARKING_LOT_GEOJSON_MODULES,
 } from "./data";
 import createDropdownGroups from "./dropdownGroups";
+import formatScorecard from "./scorecard";
 
 export default async function initApp(): Promise<void> {
   await bootstrapApp({
@@ -15,6 +16,7 @@ export default async function initApp(): Promise<void> {
       parkingLots: PARKING_LOT_GEOJSON_MODULES,
     },
     dropdownGroups: createDropdownGroups(CITY_STATS_DATA),
+    scorecardFormatter: formatScorecard,
     initialCity: "atlanta-ga",
   });
 }
