+++ packages/shared/src/js/city-ui/scorecard.ts
@@ -2,43 +2,14 @@ import type { CityEntryCollection, CityStats } from "../model/types";
 import Observable from "../state/Observable";
 import { ViewStateObservable } from "../state/ViewState";
 
-function generateScorecard(stats: CityStats): string {
-  let header = `
-      <h1 class="scorecard-title">Parking lots in ${stats.name}</h1>
-      <p>${stats.percentage} of the central city is off-street parking</p>
-      `;
+export interface ScorecardValues {
+  header: string;
+  listEntries: string[];
+}
 
-  if ("contribution" in stats) {
-    header += `<div class="community-contribution-warning">
-      <i class="fa-solid fa-triangle-exclamation" aria-hidden="true"></i> 
-      Community-maintained map
-    </div>`;
-  }
+export type ScorecardFormatter = (stats: CityStats) => ScorecardValues;
 
-  const listEntries = [];
-  if (stats.parkingScore) {
-    listEntries.push(
-      `${stats.parkingScore}/100 parking score (lower is better)`,
-    );
-  }
-  listEntries.push(`City type: ${stats.cityType}`);
-  listEntries.push(`${stats.population} residents - city proper`);
-  listEntries.push(
-    `${stats.urbanizedAreaPopulation} residents - urbanized area`,
-  );
-
-  let reformsLine = `Parking reforms ${stats.reforms}`;
-  if (stats.url) {
-    reformsLine += ` (<a class="external-link" title="view parking reform details" href="${stats.url}" target="_blank">details <i aria-hidden="true" class="fa-solid fa-arrow-right"></i></a>)`;
-  }
-  listEntries.push(reformsLine);
-
-  if (stats.contribution) {
-    listEntries.push(
-      `<a href="mailto:${stats.contribution}">Email data maintainer</a>`,
-    );
-  }
-
+function generateScorecard(values: ScorecardValues): string {
   const accordion = `<div class="scorecard-accordion">
       <button
         class="scorecard-accordion-toggle"
@@ -58,13 +29,12 @@ function generateScorecard(stats: CityStats): string {
         hidden
       >
         <ul>
-        ${listEntries.map((e) => `<li>${e}</li>`).join("\n")}
+        ${values.listEntries.map((e) => `<li>${e}</li>`).join("\n")}
         </ul>
       </section>
     </div>
   `;
-
-  return header + accordion;
+  return values.header + accordion;
 }
 
 function updateAccordionUI(expanded: boolean): void {
@@ -105,11 +75,14 @@ function initAccordion(): void {
 export default function subscribeScorecard(
   viewState: ViewStateObservable,
   cityEntries: CityEntryCollection,
+  scorecardFormatter: ScorecardFormatter,
 ): void {
   viewState.subscribe(({ cityId }) => {
     const scorecardContainer = document.querySelector(".scorecard-container");
     if (!scorecardContainer) return;
-    scorecardContainer.innerHTML = generateScorecard(cityEntries[cityId].stats);
+    scorecardContainer.innerHTML = generateScorecard(
+      scorecardFormatter(cityEntries[cityId].stats),
+    );
   }, "generate scorecard");
 
   // Also set up the accordion UI. It doesn't depend on globalState, so only
